requirements.txt
torch==1.9.0
torchvision==0.10.0
numpy==1.20.3
pandas==1.3.3
pillow==8.3.1
tqdm==4.62.2
matplotlib==3.4.3
scikit-learn==0.24.2
albumentations==1.0.3 